Vegetation fixes

+++ valen-game-darkages/src/main/java/be/twofold/valen/game/darkages/reader/vegetation/VegetationSurface.java
@@ -7,12 +7,12 @@ import java.util.*;
 
 public record VegetationSurface(
     String materialName,
-    float unknown,
+    int unknown,
     List<VegetationLod> lods
 ) {
     public static VegetationSurface read(DataSource source, int numLods) throws IOException {
         var materialName = source.readPString();
-        var unknown = source.readFloat();
+        var unknown = source.readInt();
         var lods = source.readObjects(numLods, VegetationLod::read);
 
         return new VegetationSurface(
